+++ app/Http/Controllers/UsuarioController.php
@@ -6,6 +6,8 @@ use App\Usuario;
 use App\Rol;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
+
 
 
 class UsuarioController extends Controller
@@ -18,7 +20,11 @@ class UsuarioController extends Controller
     public function index()
     {
         //
-        $usuarios  = Usuario::paginate(10);
+        if ( Auth::user()->hasRole(['Programador', 'Administrador']) ) {
+            $usuarios  = Usuario::paginate(10);
+        } else {
+            $usuarios  = Usuario::where('id', Auth::id())->paginate(10);
+        }
 
 		return View('Usuario.index', compact(['usuarios']));
     }
@@ -31,9 +37,13 @@ class UsuarioController extends Controller
     public function create()
     {
         //
-        $roles = Rol::all();
+        if ( Auth::user()->hasRole(['Programador', 'Administrador']) ) {
+            $roles = Rol::all();
 
-        return View('Usuario.create', compact(['roles']));
+            return View('Usuario.create', compact(['roles']));
+        } else {
+            abort(401, 'No Tiene permiso de acceder a la creacion de Usuarios');
+        }
 
     }
 
@@ -71,8 +81,18 @@ class UsuarioController extends Controller
     public function show(Usuario $usuario)
     {
         //
-        $roles = Rol::all();
-        return View('Usuario.show', compact(['usuario', 'roles']));
+        if ( Auth::user()->hasRole(['Programador', 'Administrador']) ) {
+            $roles = Rol::all();
+            return View('Usuario.show', compact(['usuario', 'roles']));
+        } else {
+            if ($usuario->id == Auth::id()) {
+                $roles = Rol::all();
+                return View('Usuario.show', compact(['usuario', 'roles']));
+            }else{
+                abort(401, 'No Tiene permiso de acceder a los detalles de este usuario');
+            }
+        }
+        
 
     }
 
@@ -85,9 +105,17 @@ class UsuarioController extends Controller
     public function edit(Usuario $usuario)
     {
         //
-        $roles = Rol::all();
-        return View('Usuario.edit', compact(['usuario', 'roles']));
-
+        if ( Auth::user()->hasRole(['Programador', 'Administrador']) ) {
+            $roles = Rol::all();
+            return View('Usuario.edit', compact(['usuario', 'roles']));
+        } else {
+            if ($usuario->id == Auth::id()) {
+                $roles = Rol::all();
+                return View('Usuario.edit', compact(['usuario', 'roles']));
+            }else{
+                abort(401, 'No Tiene permiso de editar este usuario');
+            }
+        }
     }
 
     /**
@@ -123,9 +151,19 @@ class UsuarioController extends Controller
      */
     public function destroy(Usuario $usuario)
     {
-        //
-        $usuario->delete();
+        
+        if ( Auth::user()->hasRole(['Programador', 'Administrador']) ) {
+            $usuario->delete();
 
-        return redirect()->route('usuarios.index');
+            return redirect()->route('usuarios.index');
+        } else {
+            if ($usuario->id == Auth::id()) {
+                $usuario->delete();
+
+                return redirect()->route('usuarios.index');
+            }else{
+                abort(401, 'No Tiene permiso de eliminar este usuario');
+            }
+        }
     }
 }
